// corpu_frontend/src/app/admin/model/TblEvents.ts
export class TblEvents {
    id: string;
    kelas: string;
    tahun: string;
    name: string;
    startDate: Date;
    endDate: Date;
    batch: number;
    groupEvent: string;
    event: string;
    method: string;
    internal: boolean;
    open: boolean;
    quota: number;
    totalTime: number;
    provider: string;
    place: string;
    location: string;
    pic: string;
    cost: number;
    result: number;
}